agentplanner/agentplanner/cloud.py
import os
import boto3  # AWS
from azure.storage.blob import BlobServiceClient  # Azure
from google.cloud import storage  # Google Cloud
from simple_salesforce import Salesforce  # Salesforce
from ibm_boto3 import resource  # IBM Cloud 
from heroku3 import from_key  # Heroku (through simple file storage or use external addons)
from aliyunsdkcore.client import AcsClient  # Alibaba Cloud

# AWS S3 Client Setup
def aws_s3_push(bucket_name, file_path, object_name):
    s3_client = boto3.client('s3')
    s3_client.upload_file(file_path, bucket_name, object_name)
    print(f"Uploaded {file_path} to S3 bucket {bucket_name} as {object_name}.")

def aws_s3_pull(bucket_name, object_name, file_path):
    s3_client = boto3.client('s3')
    s3_client.download_file(bucket_name, object_name, file_path)
    print(f"Downloaded {object_name} from S3 bucket {bucket_name} to {file_path}.")

# Azure Blob Storage 
def azure_blob_push(container_name, file_path, blob_name):
    connection_string = "your_connection_string_here"
    blob_service_client = BlobServiceClient.from_connection_string(connection_string)
    container_client = blob_service_client.get_container_client(container_name)
    with open(file_path, "rb") as data:
        container_client.upload_blob(blob_name, data)
    print(f"Uploaded {file_path} to Azure Blob Storage container {container_name} as {blob_name}.")

def azure_blob_pull(container_name, blob_name, file_path):
    connection_string = "your_connection_string_here"
    blob_service_client = BlobServiceClient.from_connection_string(connection_string)
    container_client = blob_service_client.get_container_client(container_name)
    with open(file_path, "wb") as download_file:
        download_file.write(container_client.download_blob(blob_name).readall())
    print(f"Downloaded {blob_name} from Azure Blob Storage container {container_name} to {file_path}.")

# Google Cloud Storage Client Setup
def google_cloud_push(bucket_name, file_path, blob_name):
    storage_client = storage.Client()
    bucket = storage_client.bucket(bucket_name)
    blob = bucket.blob(blob_name)
    blob.upload_from_filename(file_path)
    print(f"Uploaded {file_path} to Google Cloud Storage bucket {bucket_name} as {blob_name}.")

def google_cloud_pull(bucket_name, blob_name, file_path):
    storage_client = storage.Client()
    bucket = storage_client.bucket(bucket_name)
    blob = bucket.blob(blob_name)
    blob.download_to_filename(file_path)
    print(f"Downloaded {blob_name} from Google Cloud Storage bucket {bucket_name} to {file_path}.")

# IBM Cloud
def ibm_cloud_push(bucket_name, file_path, object_name):
    cos = resource("s3", endpoint_url="https://s3-api.us-geo.objectstorage.softlayer.net", aws_access_key_id="your_aws_access_key", aws_secret_access_key="your_aws_secret_key")
    cos.Bucket(bucket_name).upload_file(file_path, object_name)
    print(f"Uploaded {file_path} to IBM Cloud Object Storage bucket {bucket_name} as {object_name}.")

def ibm_cloud_pull(bucket_name, object_name, file_path):
    cos = resource("s3", endpoint_url="https://s3-api.us-geo.objectstorage.softlayer.net", aws_access_key_id="your_aws_access_key", aws_secret_access_key="your_aws_secret_key")
    cos.Bucket(bucket_name).download_file(object_name, file_path)
    print(f"Downloaded {object_name} from IBM Cloud Object Storage bucket {bucket_name} to {file_path}.")

# Salesforce 
def salesforce_push(file_path):
    sf = Salesforce(username="your_username", password="your_password", security_token="your_token")
    with open(file_path, "rb") as file:
        file_content = file.read()
        sf.attachments.create({'ParentId': 'record_id_here', 'Name': 'file_name', 'Body': file_content})
    print(f"Uploaded {file_path} to Salesforce.")

def salesforce_pull(attachment_id, file_path):
    sf = Salesforce(username="your_username", password="your_password", security_token="your_token")
    attachment = sf.Attachment.get(attachment_id)
    with open(file_path, "wb") as file:
        file.write(attachment['Body'])
    print(f"Downloaded attachment from Salesforce to {file_path}.")

# Heroku 
def heroku_push(file_path, addon_name="your_addon"):
    heroku_app = from_key("your_heroku_api_key")
    storage = heroku_app.addons()[addon_name].storage()
    with open(file_path, 'rb') as file:
        storage.upload(file)
    print(f"Uploaded {file_path} to Heroku addon storage.")

def heroku_pull(file_path, addon_name="your_addon"):
    heroku_app = from_key("your_heroku_api_key")
    storage = heroku_app.addons()[addon_name].storage()
    with open(file_path, 'wb') as file:
        storage.download(file)
    print(f"Downloaded file from Heroku storage to {file_path}.")
    
    
# Alibaba Cloud OSS Client Setup
def alibaba_cloud_push(bucket_name, file_path, object_name):
    client = AcsClient("<your-access-key-id>", "<your-access-key-secret>", "<your-region-id>")
    # Add Alibaba Cloud OSS upload logic using SDK
    print(f"Uploaded {file_path} to Alibaba Cloud OSS bucket {bucket_name} as {object_name}.")

def alibaba_cloud_pull(bucket_name, object_name, file_path):
    client = AcsClient("<your-access-key-id>", "<your-access-key-secret>", "<your-region-id>")
    # Add Alibaba Cloud OSS download logic using SDK
    print(f"Downloaded {object_name} from Alibaba Cloud OSS bucket {bucket_name} to {file_path}.")
